power_outage_ETL/requirements.txt
pandas
numpy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
urllib3
selenium
webdriver-manager
